companyImage

+++ src/main/java/pi/dev/realestate/repositories/CompanyImageRepository.java
@@ -2,9 +2,8 @@ package pi.dev.realestate.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import pi.dev.realestate.entities.BadWords;
 import pi.dev.realestate.entities.CompanyImage;
 
 @Repository
-public interface CompanyImageRepository extends JpaRepository<CompanyImage, Integer> {
+public interface CompanyImageRepository extends JpaRepository<CompanyImage, Long> {
 }
